Fix form feeds in received SMS messages.
Google Voice SMS sends form feeds between lines instead of newlines.
GMail SMS also puts a form feed before the appended email address.
Fixes these characters showing up as the default glyph.

This is changed in four places: the conversation list, the thread list,
the notification ticker and the actual notification text.

I'm not sure if there's a better way to do this.  Changing the GSM
conversion or otherwise changing how it's stored in the database is
wrong; that should be a clean 1:1 conversion.  This is simple enough.

I don't know why GV uses FF instead of newlines between lines.